v3.8.0（2021.01.24） 1. 新增通用功能‘版本迭代混淆’，过审后迭代更新，沿用上一次（也可以任意选择版本）混淆记录增量混淆，保持版本连续性，模拟正常开发。 a. 优势：只要维护源码，正常开发流程，避免在已过包的混淆项目上继续开发，做到开发和混淆同步且各自独立 b. 目前80%的主要功能均支持更新混淆，后续将剩余补上 2. 修复OC[重命名多语言]，项目中没有多语言的时候出现的bug 3. 修复V3.6.0版本因为调换OC[插入文件]执行顺序，导致小概率出现插入文件名重复的bug 4. 修复测试工程全部勾选情况下，小概率出现C++漏改一个了方法名 5. 修复OC[插入属性]百分比控制为100%时，属性名和类名同名极端情况，导致类名被改的bug

+++ confuse_test/confuse_test/cpp/function/CppFunction.cpp
@@ -135,10 +135,10 @@ void CppFunction4::test() {
 }
 
 // MARK: - CppFunction5
-void CppFunction5::sf1(float v) {
+void CppFunction5::sf1(double v) {
     Function_Log
 }
-int CppFunction5::sf1(int v1, float v2) {
+int CppFunction5::sf1(int v1, double v2) {
     Function_Log return v1;
 }
 
@@ -151,10 +151,10 @@ void CppFunction5::test() {
 
 namespace custom {
     // MARK: - CppFunction5
-    void CppFunction5::sf1(float v) {
+    void CppFunction5::sf1(double v) {
         Function_Log
     }
-    int CppFunction5::sf1(int v1, float v2) {
+    int CppFunction5::sf1(int v1, double v2) {
         Function_Log return v1;
     }
 
@@ -166,11 +166,11 @@ namespace custom {
     }
 }  // namespace custom
 
-void CppFunction6::f1(float v) {
+void CppFunction6::f1(double v) {
     Function_Log
 }
 
-int CppFunction6::f1(int v1, float v2) {
+int CppFunction6::f1(int v1, double v2) {
     Function_Log return v1;
 }
